ATTN: disabling build of Burlap
The burlap source code assumes that the main data structures are POD
types, which makes it difficult to use c++ containters in them. I
think that these abstractions are necessary for robust code, and given
that burlap will soon be replaced by a lua wrapper library (developed
in another branch), I see no reason to keep it around.